Add feature to retrieve all users
Added new classes for retrieving all users from the to-do list application. This includes a new use case 'GetUsersUseCase', a corresponding web controller 'GetUsersController', and a service 'UserService'. Also included is a 'User' model and an associated Thymeleaf template for displaying the users. This provides the functionality for user retrieval which is essential for user management activities in the application.