Add user loading functionality
Implemented the ability to load users from a persistent storage. Added model classes, 'User.java', 'UserApiEntity.java', and an interface for loading users, 'LoadUsersPort.java'. Also, included User service and controller classes which handle the loading operation and make it available for the application. The reason behind this change is to provide the users data to the application from a persistent storage.